EventLabelViewクラスに切り分け
